requirements.txt
Flask==0.10.1
pandas==0.19.2
scikit-learn==0.18.1
Scipy==0.19.0
requests==2.13.0 
python-socketio==1.7.4
eventlet==0.20.1
python-twitter==3.2.1